requirements.txt
pyyaml
snowflake-cli-labs